service layer: creation and implementation-RUD
