--- src/main/java/com/santox/crudDEmo/service/studentService.java
@@ -1,7 +1,0 @@
-package com.santox.crudDEmo.service;
-
-import com.santox.crudDEmo.entity.Student;
-
-public interface studentService {
-    Student save(Student student);
-}
