--- src/main/java/com/santox/crudDEmo/service/studentServiceImpl.java
@@ -1,10 +1,0 @@
-package com.santox.crudDEmo.service;
-
-import com.santox.crudDEmo.entity.Student;
-
-public class studentServiceImpl implements studentService{
-    @Override
-    public Student save(Student student) {
-        return null;
-    }
-}
